requirements.txt
tensorflow-gpu==2.9.0
matplotlib==3.5.3
mpire==2.6.0
pandas==1.4.4
PyYAML==6.0
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.0
sklearn==0.0
spotpy==1.6.1
tensorboard==2.9.0
tensorflow-probability==0.14.0
tqdm==4.64.1
ipython==8.5.0
